Clean warnings (comparison signed and unsigned)

+++ escher/src/text_field.cpp
@@ -91,7 +91,7 @@ bool TextField::ContentView::insertTextAtLocation(const char * text, int locatio
     m_draftTextBuffer[location+textSize-1] = text[textSize-1];
   }
   m_currentDraftTextLength += textSize;
-  for (int i = 0; i < m_currentDraftTextLength; i++) {
+  for (size_t i = 0; i < m_currentDraftTextLength; i++) {
     if (m_draftTextBuffer[i] == '\n') {
       m_draftTextBuffer[i] = 0;
       m_currentDraftTextLength = i;
